requirement.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-pymemcache==1.0.0
idna==2.10
pymemcache==3.4.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
